refactor: repository의 magic number 제거
