+++ src/main/java/nextstep/reservations/domain/repository/reservation/JdbcTemplateReservationRepository.java
@@ -55,10 +55,8 @@ public class JdbcTemplateReservationRepository implements ReservationRepository{
     }
 
     @Override
-    public void remove(final Long id) {
-        int count = jdbcTemplate.update(ReservationQuery.REMOVE_BY_ID.get(), id);
-
-        if (count == 0) throw new NoSuchReservationException();
+    public int remove(final Long id) {
+        return jdbcTemplate.update(ReservationQuery.REMOVE_BY_ID.get(), id);
     }
 
     private static RowMapper<Reservation> getReservationRowMapper() {
