+++ src/main/java/nextstep/reservations/domain/service/reservation/ReservationService.java
@@ -3,6 +3,7 @@ package nextstep.reservations.domain.service.reservation;
 import nextstep.reservations.domain.repository.reservation.ReservationRepository;
 import nextstep.reservations.dto.reservation.ReservationRequestDto;
 import nextstep.reservations.dto.reservation.ReservationResponseDto;
+import nextstep.reservations.exceptions.reservation.exception.NoSuchReservationException;
 import nextstep.reservations.util.mapper.ReservationMapper;
 import org.springframework.stereotype.Service;
 
@@ -25,6 +26,8 @@ public class ReservationService {
     }
 
     public void deleteReservation(final Long id) {
-        reservationRepository.remove(id);
+        int removeCount = reservationRepository.remove(id);
+
+        if (removeCount == 0) throw new NoSuchReservationException();
     }
 }
