+++ src/main/java/nextstep/reservations/exceptions/theme/ThemeControllerAdvice.java
@@ -1,7 +1,7 @@
 package nextstep.reservations.exceptions.theme;
 
-import nextstep.reservations.domain.controller.reservation.ReservationController;
-import nextstep.reservations.domain.controller.theme.ThemeController;
+import nextstep.reservations.controller.reservation.ReservationController;
+import nextstep.reservations.controller.theme.ThemeController;
 import nextstep.reservations.exceptions.theme.exception.DuplicateThemeException;
 import nextstep.reservations.exceptions.theme.exception.NoSuchThemeException;
 import org.springframework.http.ResponseEntity;
